Add profiling to various modules using the Profiler class.
I also had to move the get_cuda_devices method, since i would do circular imports otherwise
