+++ src/qttools/greens_function_solver/rgf.py
@@ -3,9 +3,13 @@
 from qttools import NDArray, xp
 from qttools.datastructures.dsbsparse import DSBSparse
 from qttools.greens_function_solver.solver import GFSolver, OBCBlocks
+from qttools.profiling import Profiler, decorate_methods
 from qttools.utils.solvers_utils import get_batches
 
+profiler = Profiler()
 
+
+@decorate_methods(profiler.profile(level="api"), exclude=["__init__"])
 class RGF(GFSolver):
     """Selected inversion solver based on the Schur complement.
