+++ src/qttools/kernels/linalg/eigvalsh.py
@@ -1,9 +1,13 @@
 import numpy as np
 
 from qttools import NDArray, xp
+from qttools.profiling import Profiler
 from qttools.utils.gpu_utils import get_any_location, get_array_module_name
 
+profiler = Profiler()
 
+
+@profiler.profile(level="debug")
 def _eigvalsh(
     A: NDArray,
     B: NDArray,
@@ -38,6 +42,7 @@ def _eigvalsh(
     return w
 
 
+@profiler.profile(level="api")
 def eigvalsh(
     A: NDArray,
     B: NDArray | None = None,
