+++ src/qttools/kernels/linalg/qr.py
@@ -4,9 +4,13 @@ import numba as nb
 import numpy as np
 
 from qttools import NDArray, xp
+from qttools.profiling import Profiler
 from qttools.utils.gpu_utils import get_any_location, get_array_module_name
 
+profiler = Profiler()
 
+
+@profiler.profile(level="debug")
 @nb.njit(parallel=True, cache=True, no_rewrites=True)
 def _qr_numba(
     A: NDArray,
@@ -46,6 +50,7 @@ def _qr_numba(
     return q, r
 
 
+@profiler.profile(level="api")
 def qr(
     A: NDArray,
     compute_module: str = "numpy",
