+++ src/qttools/kernels/numba/dsbcsr.py
@@ -4,7 +4,12 @@ import numba as nb
 import numpy as np
 from numpy.typing import NDArray
 
+from qttools.profiling import Profiler
 
+profiler = Profiler()
+
+
+@profiler.profile(level="debug")
 @nb.njit(parallel=True, cache=True)
 def _find_bcoords(block_offsets: NDArray, rows: NDArray, cols: NDArray) -> NDArray:
     """Finds the block coordinates of the given rows and columns.
@@ -38,6 +43,7 @@ def _find_bcoords(block_offsets: NDArray, rows: NDArray, cols: NDArray) -> NDArr
     return brows, bcols
 
 
+@profiler.profile(level="debug")
 @nb.njit(parallel=True, cache=True)
 def _find_block_inds(
     brows: NDArray,
@@ -103,6 +109,7 @@ def _find_block_inds(
     return inds[valid], mask_inds[valid]
 
 
+@profiler.profile(level="api")
 def find_inds(
     rowptr_map: dict[tuple, NDArray],
     block_offsets: NDArray,
@@ -162,6 +169,7 @@ def find_inds(
     return np.array(inds, dtype=int), np.array(value_inds, dtype=int)
 
 
+@profiler.profile(level="api")
 @nb.njit(parallel=True, cache=True)
 def densify_block(
     block: NDArray,
@@ -191,6 +199,7 @@ def densify_block(
         block[..., i, cols] = data[..., rowptr[i] : rowptr[i + 1]]
 
 
+@profiler.profile(level="api")
 @nb.njit(parallel=True, cache=True)
 def sparsify_block(
     block: NDArray,
@@ -220,7 +229,9 @@ def sparsify_block(
         data[..., rowptr[i] : rowptr[i + 1]] = block[..., i, cols]
 
 
+# TODO: Check why the jit is uncommented here.
 # @nb.njit(parallel=True, cache=True)
+@profiler.profile(level="debug")
 def _compute_rowptr_map_kernel(
     coo_rows: NDArray, coo_cols: NDArray, block_sizes: NDArray
 ) -> nb.typed.Dict:
@@ -302,6 +313,7 @@ def _compute_rowptr_map_kernel(
     return sort_index, rowptr_map
 
 
+@profiler.profile(level="api")
 def compute_rowptr_map(
     coo_rows: NDArray, coo_cols: NDArray, block_sizes: NDArray
 ) -> dict:
