+++ src/qttools/lyapunov/spectral.py
@@ -6,6 +6,9 @@ from qttools import NDArray, xp
 from qttools.kernels import linalg
 from qttools.lyapunov.lyapunov import LyapunovSolver
 from qttools.lyapunov.utils import system_reduction
+from qttools.profiling import Profiler
+
+profiler = Profiler()
 
 
 class Spectral(LyapunovSolver):
@@ -38,6 +41,7 @@ class Spectral(LyapunovSolver):
         self.eig_compute_location = eig_compute_location
         self.reduce_sparsity = reduce_sparsity
 
+    @profiler.profile(level="debug")
     def _solve(
         self,
         a: NDArray,
@@ -96,6 +100,7 @@ class Spectral(LyapunovSolver):
 
         return x_ref
 
+    @profiler.profile(level="api")
     def __call__(
         self,
         a: NDArray,
