+++ src/qttools/lyapunov/utils.py
@@ -1,6 +1,10 @@
 from qttools import NDArray, xp
+from qttools.profiling import Profiler
+
+profiler = Profiler()
 
 
+@profiler.profile(level="debug")
 def _system_reduction_rows(a: NDArray, q: NDArray, solve, rows_to_reduce):
     """Reduces the system by rows of A that are all zero.
 
@@ -35,6 +39,7 @@ def _system_reduction_rows(a: NDArray, q: NDArray, solve, rows_to_reduce):
     return x
 
 
+@profiler.profile(level="debug")
 def _system_reduction_cols(
     a: NDArray,
     q: NDArray,
@@ -74,6 +79,7 @@ def _system_reduction_cols(
     return x
 
 
+@profiler.profile(level="debug")
 def system_reduction(
     a: NDArray,
     q: NDArray,
