+++ src/qttools/obc/obc.py
@@ -6,6 +6,9 @@ from mpi4py import MPI
 from mpi4py.MPI import COMM_WORLD as comm
 
 from qttools import NDArray, xp
+from qttools.profiling import Profiler
+
+profiler = Profiler()
 
 
 class OBCSolver(ABC):
@@ -79,6 +82,7 @@ class OBCMemoizer:
         self.convergence_tol = convergence_tol
         self._cache = {}
 
+    @profiler.profile(level="debug")
     def _call_with_cache(
         self,
         a_ii: NDArray,
@@ -117,6 +121,7 @@ class OBCMemoizer:
         self._cache[contact] = out.copy()
         return None
 
+    @profiler.profile(level="api")
     def __call__(
         self,
         a_ii: NDArray,
