+++ src/qttools/utils/solvers_utils.py
@@ -1,9 +1,13 @@
 # Copyright (c) 2024 ETH Zurich and the authors of the qttools package.
 
 from qttools import NDArray, xp
+from qttools.profiling import Profiler
 from qttools.utils.mpi_utils import get_section_sizes
 
+profiler = Profiler()
 
+
+@profiler.profile(level="debug")
 def get_batches(num_sections: int, max_batch_size: int) -> tuple[list, NDArray]:
     """Computes the number of batches and their sizes.
